requirements.txt
certifi==2020.6.20
chardet==3.0.4
enum34==1.1.10
future==0.18.2
idna==2.10
numpy==1.19.1
opencv-python==4.3.0.36
Pillow==7.2.0
pytorch-ignite==0.1.0
requests==2.24.0
torch==1.6.0
torchvision==0.7.0
tqdm==4.48.1
urllib3==1.25.10
wrapt==1.12.1
